pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "p4"
version = "0.1.0"
description = "AI-Powered Power Plant Performance Predictor"
requires-python = ">=3.10"
dependencies = [
    "pyside6>=6.6",
    "sqlalchemy>=2.0",
    "pyodbc>=5.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "torch>=2.0",
    "onnxruntime>=1.16",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "numpy>=1.26",
    "pandas>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
p4 = "p4.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
